update diffusion models | bf16

+++ src/unitorch/cli/models/diffusers/modeling_stable_xl.py
@@ -196,10 +196,7 @@ class StableXLForText2ImageGeneration(_StableXLForText2ImageGeneration):
             )
         return inst
 
-    @autocast(
-        device_type=("cuda" if torch.cuda.is_available() else "cpu"),
-        dtype=torch.bfloat16,
-    )
+    @autocast(device_type=("cuda" if torch.cuda.is_available() else "cpu"))
     def forward(
         self,
         input_ids: torch.Tensor,
@@ -220,10 +217,7 @@ class StableXLForText2ImageGeneration(_StableXLForText2ImageGeneration):
         return LossOutputs(loss=loss)
 
     @add_default_section_for_function("core/model/diffusers/text2image/stable_xl")
-    @autocast(
-        device_type=("cuda" if torch.cuda.is_available() else "cpu"),
-        dtype=torch.bfloat16,
-    )
+    @autocast(device_type=("cuda" if torch.cuda.is_available() else "cpu"))
     def generate(
         self,
         input_ids: torch.Tensor,
@@ -431,10 +425,7 @@ class StableXLForImage2ImageGeneration(_StableXLForImage2ImageGeneration):
         raise NotImplementedError
 
     @add_default_section_for_function("core/model/diffusers/image2image/stable_xl")
-    @autocast(
-        device_type=("cuda" if torch.cuda.is_available() else "cpu"),
-        dtype=torch.bfloat16,
-    )
+    @autocast(device_type=("cuda" if torch.cuda.is_available() else "cpu"))
     def generate(
         self,
         input_ids: torch.Tensor,
@@ -636,10 +627,7 @@ class StableXLForImageInpainting(_StableXLForImageInpainting):
             )
         return inst
 
-    @autocast(
-        device_type=("cuda" if torch.cuda.is_available() else "cpu"),
-        dtype=torch.bfloat16,
-    )
+    @autocast(device_type=("cuda" if torch.cuda.is_available() else "cpu"))
     def forward(
         self,
         input_ids: torch.Tensor,
@@ -662,10 +650,7 @@ class StableXLForImageInpainting(_StableXLForImageInpainting):
         return LossOutputs(loss=loss)
 
     @add_default_section_for_function("core/model/diffusers/inpainting/stable_xl")
-    @autocast(
-        device_type=("cuda" if torch.cuda.is_available() else "cpu"),
-        dtype=torch.bfloat16,
-    )
+    @autocast(device_type=("cuda" if torch.cuda.is_available() else "cpu"))
     def generate(
         self,
         input_ids: torch.Tensor,
